Run mypy on core library (#2665)
* Run mypy on core library

* Revert to relpath

* More mypy

* More mypy

* More mypy and fix

* Continute with plotly codebase

* Line plot custom markers: customize inner and outer colors

* Mypy more plotly code

* Use generic KeyT and ValueT in line graph code

* Ignore dot files in picard test

* More mypy in plotly

* Fix tests

* Fix ZeroDivisionError

* More mypy

* More mypy

* More mypy

* Add mypy into precommit

* Pin mypy in precommit

* Add mypy CI job

* Remove mypy from pre-commit

* Move mypy to dev installation

* Document mypy

* Ignore missing IPython stubs

* Static-type and refactor barplot cats, add more tests

* Static-type picard

* Mypy bcftools, add tests for barplot: single cats, mupltiple datasets

* Mypy kallisto

* Mypy samtools

* Mypy profile_runtime

* Fix star module and mypyise

* Mypy setup.py

+++ multiqc/core/exec_modules.py
@@ -5,6 +5,7 @@ import time
 import traceback
 import tracemalloc
 from importlib.metadata import EntryPoint
+from pathlib import Path
 from typing import Dict, Union, Callable, List
 
 import rich
@@ -63,8 +64,8 @@ def exec_modules(
         try:
             entry_point: EntryPoint = config.avail_modules[this_module]
             module_initializer: Callable[[], Union[BaseMultiqcModule, List[BaseMultiqcModule]]] = entry_point.load()
-            module_initializer.mod_cust_config = mod_cust_config
-            module_initializer.mod_id = this_module
+            setattr(module_initializer, "mod_cust_config", mod_cust_config)
+            setattr(module_initializer, "mod_id", this_module)
 
             # *********************************************
             # RUN MODULE. Heavy part. Run module logic to parse logs and prepare plot data.
@@ -112,25 +113,26 @@ def exec_modules(
 
             # Flag the error, but carry on
             class CustomTraceback:
+                type, value, traceback = sys.exc_info()
+
                 def __rich_console__(self, console: rich.console.Console, options: rich.console.ConsoleOptions):
-                    sys_tb = sys.exc_info()
-                    issue_url = "https://github.com/MultiQC/MultiQC/issues/new?template=bug_report.md&title={}%20module%20-%20{}".format(
-                        this_module, sys_tb[0].__name__
+                    issue_url = f"https://github.com/MultiQC/MultiQC/issues/new?template=bug_report.md&title={this_module}%20module%20-%20{type.__name__}"
+                    err_msg = (
+                        f"Please copy this log and report it at [bright_blue][link={issue_url}]"
+                        f"https://github.com/MultiQC/MultiQC/issues[/link][/] \n"
+                        f"[bold underline]Please attach a file that triggers the error.[/] "
                     )
-                    yield (
-                        "Please copy this log and report it at [bright_blue][link={}]https://github.com/MultiQC/MultiQC/issues[/link][/] \n"
-                        "[bold underline]Please attach a file that triggers the error.[/] The last file found was: [green]{}[/]\n".format(
-                            issue_url, report.last_found_file
-                        )
-                    )
+                    if report.last_found_file:
+                        err_msg += f"The last file found was: [green]{report.last_found_file}[/]\n"
+
+                    yield err_msg
                     yield Syntax(traceback.format_exc(), "python")
 
                 def __rich_measure__(self, console: rich.console.Console, options: rich.console.ConsoleOptions):
                     tb_width = max([len(line) for line in traceback.format_exc().split("\n")])
-                    try:
-                        log_width = 71 + len(report.last_found_file)
-                    except TypeError:
-                        log_width = 71
+                    log_width = 71
+                    if report.last_found_file:
+                        log_width += len(report.last_found_file)
                     panel_width = max(tb_width, log_width)
                     return rich.console.Measurement(panel_width, panel_width)
 
@@ -156,7 +158,7 @@ def exec_modules(
             # Exit code 1 for CI failures etc
             sys_exit_code = 1
 
-        report.runtimes["mods"][mod_names[mod_idx]] = time.time() - mod_starttime
+        report.runtimes.mods[mod_names[mod_idx]] = time.time() - mod_starttime
         if config.profile_memory:
             mem_current, mem_peak = tracemalloc.get_traced_memory()
             tracemalloc.stop()
@@ -166,9 +168,9 @@ def exec_modules(
                 f"{this_module}: memory change: {mem_current:,d}b, peak during module execution: {mem_peak:,d}b"
             )
         if config.profile_runtime:
-            logger.warning(f"{this_module}: module run time: {report.runtimes['mods'][mod_names[mod_idx]]:.2f}s")
+            logger.warning(f"{this_module}: module run time: {report.runtimes.mods[mod_names[mod_idx]]:.2f}s")
 
-    report.runtimes["total_mods"] = time.time() - total_mods_starttime
+    report.runtimes.total_mods = time.time() - total_mods_starttime
 
     # Again, if config.require_logs is set, check if for all explicitly requested
     # modules samples were found.
@@ -181,7 +183,7 @@ def exec_modules(
     # Did we find anything?
     if len(report.modules) == 0:
         logger.warning("No analysis results found. Cleaning up…")
-        if clean_up:
+        if clean_up and report.tmp_dir and Path(report.tmp_dir).exists():
             shutil.rmtree(report.tmp_dir)
         logger.info("MultiQC complete")
         # Exit with an error code if a module broke
