+++ multiqc/core/write_results.py
@@ -10,7 +10,7 @@ import sys
 import time
 import traceback
 from pathlib import Path
-from typing import Set, List
+from typing import Set
 
 from multiqc.base_module import Section
 from multiqc.core.file_search import include_or_exclude_modules
@@ -48,7 +48,7 @@ def write_results(clean_up=True) -> None:
 
     overwritten = _create_or_override_dirs() if config.filename != "stdout" else set()
 
-    if config.make_data_dir and config.filename != "stdout":
+    if config.make_data_dir and config.filename != "stdout" and config.data_dir:
         _write_data_files()
         logger.info(
             "Data        : {}{}".format(
@@ -62,22 +62,23 @@ def write_results(clean_up=True) -> None:
     if config.make_report:
         # Render report HTML, write to file or stdout
         _write_report()
-        if config.filename != "stdout":
-            if config.make_report:
+        if config.filename != "stdout" and config.output_fn:
+            assert isinstance(config.output_fn, str)
+            if config.make_report and config.output_fn:
                 logger.info(
                     "Report      : {}{}".format(
                         os.path.relpath(config.output_fn),
                         "   (overwritten)" if "report" in overwritten else "",
                     )
                 )
-                logger.debug(f"Full report path: {os.path.realpath(config.output_fn)}")
+                logger.debug(f"Full report path: {os.path.realpath(Path(config.output_fn))}")
             else:
                 logger.info("Report      : None")
         # Try to create a PDF if requested
         if config.make_pdf:
             _write_pdf()
 
-    if config.export_plots:
+    if config.export_plots and config.plots_dir:
         # Copy across the static plot images if requested
         _move_exported_plots()
         logger.info(
@@ -88,7 +89,7 @@ def write_results(clean_up=True) -> None:
         )
 
     # Clean up temporary directory
-    if clean_up:
+    if clean_up and report.tmp_dir and os.path.isdir(report.tmp_dir):
         shutil.rmtree(report.tmp_dir)
 
     # Zip the data directory if requested
@@ -176,45 +177,49 @@ def _create_or_override_dirs() -> Set[str]:
 
     # Check for existing reports and remove if -f was specified
     if (
-        (config.make_report and os.path.exists(config.output_fn))
-        or (config.make_data_dir and os.path.exists(config.data_dir))
-        or (config.export_plots and os.path.exists(config.plots_dir))
+        (config.make_report and isinstance(config.output_fn, str) and os.path.exists(config.output_fn))
+        or (config.make_data_dir and config.data_dir and os.path.exists(config.data_dir))
+        or (config.export_plots and config.plots_dir and os.path.exists(config.plots_dir))
     ):
         if config.force:
-            if config.make_report and os.path.exists(config.output_fn):
+            if config.make_report and isinstance(config.output_fn, str) and os.path.exists(config.output_fn):
                 overwritten.add("report")
                 os.remove(config.output_fn)
-            if config.make_data_dir and os.path.exists(config.data_dir):
+            if config.make_data_dir and config.data_dir and os.path.exists(config.data_dir):
                 overwritten.add("data_dir")
                 shutil.rmtree(config.data_dir)
-            if config.export_plots and os.path.exists(config.plots_dir):
+            if config.export_plots and config.plots_dir and os.path.exists(config.plots_dir):
                 overwritten.add("export_plots")
                 shutil.rmtree(config.plots_dir)
         else:
             # Set up the base names of the report and the data dir
             report_num = 1
-            dir_base = os.path.basename(config.data_dir)
-            plots_base = os.path.basename(config.plots_dir)
 
             # Iterate through appended numbers until we find one that's free
             while (
-                (config.make_report and os.path.exists(config.output_fn))
-                or (config.make_data_dir and os.path.exists(config.data_dir))
-                or (config.export_plots and os.path.exists(config.plots_dir))
+                (config.make_report and isinstance(config.output_fn, str) and os.path.exists(config.output_fn))
+                or (config.make_data_dir and config.data_dir and os.path.exists(config.data_dir))
+                or (config.export_plots and config.plots_dir and os.path.exists(config.plots_dir))
             ):
                 if config.make_report:
                     report_base, report_ext = os.path.splitext(config.output_fn_name)
                     config.output_fn = os.path.join(config.output_dir, f"{report_base}_{report_num}{report_ext}")
-                config.data_dir = os.path.join(config.output_dir, f"{dir_base}_{report_num}")
-                config.plots_dir = os.path.join(config.output_dir, f"{plots_base}_{report_num}")
+                if config.data_dir:
+                    dir_base = os.path.basename(config.data_dir)
+                    config.data_dir = os.path.join(config.output_dir, f"{dir_base}_{report_num}")
+                if config.plots_dir:
+                    plots_base = os.path.basename(config.plots_dir)
+                    config.plots_dir = os.path.join(config.output_dir, f"{plots_base}_{report_num}")
                 report_num += 1
-            if config.make_report:
+            if config.make_report and isinstance(config.output_fn, str):
                 config.output_fn_name = os.path.basename(config.output_fn)
-            config.data_dir_name = os.path.basename(config.data_dir)
-            config.plots_dir_name = os.path.basename(config.plots_dir)
+            if config.data_dir:
+                config.data_dir_name = os.path.basename(config.data_dir)
+            if config.plots_dir:
+                config.plots_dir_name = os.path.basename(config.plots_dir)
             logger.info("Existing reports found, adding suffix to filenames. Use '--force' to overwrite.")
 
-    if config.make_report:
+    if config.make_report and isinstance(config.output_fn, str):
         os.makedirs(os.path.dirname(config.output_fn), exist_ok=True)
 
     return overwritten
@@ -265,7 +270,7 @@ def _order_modules_and_sections():
                 section_id_order[anchor] = section_id_order[ss["after"]] + 1
             if ss.get("before") in section_id_order.keys():
                 section_id_order[anchor] = section_id_order[ss["before"]] - 1
-        sorted_ids = sorted(section_id_order, key=section_id_order.get)
+        sorted_ids = sorted(section_id_order.keys(), key=lambda k: section_id_order[k])
         report.modules = [mod for i in reversed(sorted_ids) for mod in report.modules if mod.anchor == i]
 
     # Sort the report sections if we have a config
@@ -297,7 +302,7 @@ def _order_modules_and_sections():
             # Remove module sections
             section_id_order = {s: o for s, o in section_id_order.items() if o is not False}
             # Sort the module sections
-            sorted_ids = sorted(section_id_order, key=section_id_order.get)
+            sorted_ids = sorted(section_id_order.keys(), key=lambda k: section_id_order[k])
             report.modules[midx].sections = [s for i in sorted_ids for s in mod.sections if s.anchor == i]
 
 
@@ -379,7 +384,7 @@ def _render_general_stats_table() -> None:
             "raw_data_fn": "multiqc_general_stats",
         }
         p = table.plot(report.general_stats_data, report.general_stats_headers, pconfig)
-        report.general_stats_html = p.add_to_report()
+        report.general_stats_html = p.add_to_report() if isinstance(p, Plot) else p
     else:
         config.skip_generalstats = True
 
@@ -390,6 +395,8 @@ def _write_data_files() -> None:
     """
     # Modules have run, so data directory should be complete by now. Move its contents.
     logger.debug(f"Moving data file from '{report.data_tmp_dir()}' to '{config.data_dir}'")
+    assert config.data_dir is not None
+
     shutil.copytree(
         report.data_tmp_dir(),
         config.data_dir,
@@ -402,10 +409,10 @@ def _write_data_files() -> None:
     shutil.rmtree(report.data_tmp_dir())
 
     # Write the report sources to disk
-    report.data_sources_tofile()
+    report.data_sources_tofile(config.data_dir)
 
     # Create a file with the module DOIs
-    report.dois_tofile(report.modules)
+    report.dois_tofile(config.data_dir, report.modules)
 
     # Data Export / MegaQC integration - save report data to file or send report data to an API endpoint
     if config.data_dump_file or (config.megaqc_url and config.megaqc_upload):
@@ -425,6 +432,8 @@ def _write_report():
     """
     Render and write report HTML to disk
     """
+    assert report.tmp_dir
+
     # Copy over css & js files if requested by the theme
     for mod in report.modules:
         if mod.hidden:
@@ -517,7 +526,7 @@ def _write_report():
     runtime_compression_start = time.time()
     logger.debug("Compressing plot data")
     report.plot_compressed_json = report.compress_json(report.plot_data)
-    report.runtimes["total_compression"] = time.time() - runtime_compression_start
+    report.runtimes.total_compression = time.time() - runtime_compression_start
 
     # Use jinja2 to render the template and overwrite
     report.analysis_files = [os.path.realpath(d) for d in report.analysis_files]
@@ -525,6 +534,7 @@ def _write_report():
     if config.filename == "stdout":
         print(report_output.encode("utf-8"), file=sys.stdout)
     else:
+        assert isinstance(config.output_fn, str)
         try:
             with io.open(config.output_fn, "w", encoding="utf-8") as f:
                 print(report_output, file=f)
@@ -533,15 +543,17 @@ def _write_report():
 
         # Copy over files if requested by the theme
         try:
-            for f in template_mod.copy_files:
-                fn = os.path.join(report.tmp_dir, f)
-                dest_dir = os.path.join(os.path.dirname(config.output_fn), f)
+            for copy_file in template_mod.copy_files:
+                fn = os.path.join(report.tmp_dir, copy_file)
+                dest_dir = os.path.join(os.path.dirname(config.output_fn), copy_file)
                 shutil.copytree(fn, dest_dir, dirs_exist_ok=True)
         except AttributeError:
             pass  # No files to copy
 
 
 def _write_pdf():
+    if not isinstance(config.output_fn, str):
+        return
     try:
         pdf_fn_name = config.output_fn.replace(".html", ".pdf")
         pandoc_call = [
